+++ DEPLOYMENT.md
@@ -1,0 +1,3 @@
+# Deployment
+
+Will come later multiple ways of deployment
