+++ cmd/exitcodes.go
@@ -3,4 +3,5 @@ package cmd
 const (
 	ExitSuccess      = 0
 	ExitGeneralError = 1
+	ExitUsageError   = 64 // Used incorrectly
 )
